build fixes
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@46212 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/string/strcache.cpp
@@ -280,11 +280,11 @@ typedef struct RTSTRCACHEINT
     /** The total size of all the big entries. */
     size_t                  cbBigEntries;
     /** Hash collisions. */
-    size_t                  cHashCollisions;
+    uint32_t                cHashCollisions;
     /** Secondary hash collisions. */
-    size_t                  cHashCollisions2;
+    uint32_t                cHashCollisions2;
     /** The number of inserts to compare cHashCollisions to. */
-    size_t                  cHashInserts;
+    uint32_t                cHashInserts;
     /** The number of rehashes. */
     uint32_t                cRehashes;
     /** @} */
